Added react router and routing logic
