+++ src/components/Home/About.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import '../styles/About.scss'
+import '../../styles/About.scss'
 
 export default function About() {
   return (
