+++ src/components/Home/Contact.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import "../styles/Contact.scss"
+import "../../styles/Contact.scss"
 import Footer from "./Footer"
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
