+++ src/components/Home/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "../styles/Footer.scss";
+import "../../styles/Footer.scss";
 
 export default function Footer() {
   return (
