+++ src/components/Home/Navbar.jsx
@@ -1,11 +1,11 @@
-import "../styles/Navbar.scss";
+import "../../styles/Navbar.scss";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
 export default function Navbar(props) {
 
   return (
     <div className={"navbar " + (props.openMenu && "active")}>
-      <div className="logo"><a href="#home">Wei An Bow</a></div>
+      <div className="logo"><a href="#landing">Wei An Bow</a></div>
       <div className="icons">
         <FontAwesomeIcon icon='angle-right' onClick={props.handleClick} className="exit"/>
         <a href="https://www.linkedin.com/in/wei-an-bow-0610b1217/" target='_blank'><FontAwesomeIcon icon="fa-brands fa-linkedin-in" /></a>
