+++ src/components/Project/ProjectPage.jsx
@@ -1,0 +1,9 @@
+import React from "react"
+
+export default function Project({projectData, icons}) {
+  return(
+    <>
+    <h1>Project page test</h1>
+    </>
+  )
+};
